requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
idna==2.8
lxml==4.4.1
multidict==4.5.2
requests==2.22.0
slackclient==2.2.1
urllib3==1.25.6
yarl==1.3.0
